Move to pdftk-java on debian

+++ watch.ps1
@@ -1,7 +1,5 @@
-param(
-    [String]$IN_DIR = "/files-in", 
-    [String]$OUT_DIR = "/files-out"
-)
+$IN_DIR = "/files-in"
+$OUT_DIR = "/files-out"
 
 $previous_file_name = $null
 
@@ -28,8 +26,12 @@ inotifywait -q -m -e CLOSE_WRITE $IN_DIR | Foreach-Object {
             $previous_file_name = Join-Path $IN_DIR $previous_file_name
 
             Write-Information "Executing pdftk A=$previous_file_name B=$file_name shuffle A Bend-1 output $out_file"
-            pdftk A=$previous_file_name B=$file_name shuffle A Bend-1 output $out_file
+            # todo does not ask before overwriting
+            ./pdftk A=$previous_file_name B=$file_name shuffle A Bend-1 output $out_file
 
+            if (-not $?) {
+                throw "pdftk errored; exiting."
+            }
             Remove-Item $file_name, $previous_file_name
             $previous_file_name = $null
         }
